refactor: split db/scans.ts by what it does to a scan
1233 lines covering six unrelated jobs. Split by operation rather than by
caller, since the callers already overlap heavily:

- scan-jobs.ts      create / claim / fail / discard — lifecycle, no results
- scan-persist.ts   persistScan and its claim-guarded batched inserts
- scan-list.ts      the org-scoped, keyset-paginated dashboard list
- scan-detail.ts    reading one scan back, D1 rows merged over R2 artifacts
- scan-decisions.ts publish / no-publish verdicts and their audit events
- scan-risk.ts      persisted risk-summary readers shared by the above
- d1-chunk.ts       bound-parameter sizing, previously buried mid-file

db/scans.ts stays the import path for all 27 consumers and is now a barrel
naming what lives where. No behavior change: every top-level declaration in
the original is present in exactly one of the new modules, and the full
suite (2198 tests) passes unchanged.
